requirements.txt
streamlit
python-dotenv
sqlparse
transformers
requests
huggingface_hub
sentence_transformers
tqdm